refactor: lib to libs, remove rust warning
