+++ src-tauri/src/libs/s3/get/buckets.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-use crate::lib::s3::{
+use crate::libs::s3::{
     client::client::create_client,
     utils::response_error::{create_error, ResponseError},
 };
