+++ src-tauri/src/libs/s3/get/images.rs
@@ -5,9 +5,9 @@ use itertools::Itertools;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 
-use crate::lib::s3::client::client::create_client;
-use crate::lib::s3::utils::presigned_url::get_presigned_url;
-use crate::lib::s3::utils::response_error::{create_error, ResponseError};
+use crate::libs::s3::client::client::create_client;
+use crate::libs::s3::utils::presigned_url::get_presigned_url;
+use crate::libs::s3::utils::response_error::{create_error, ResponseError};
 use std::error::Error;
 use tokio_stream::StreamExt;
