+++ src-tauri/src/libs/s3/get/movies.rs
@@ -2,9 +2,9 @@ use aws_sdk_s3::Client;
 use cached::proc_macro::once;
 use serde::{Deserialize, Serialize};
 
-use crate::lib::s3::client::client::create_client;
-use crate::lib::s3::utils::presigned_url::get_presigned_url;
-use crate::lib::s3::utils::response_error::{create_error, ResponseError};
+use crate::libs::s3::client::client::create_client;
+use crate::libs::s3::utils::presigned_url::get_presigned_url;
+use crate::libs::s3::utils::response_error::{create_error, ResponseError};
 use aws_sdk_s3::types::Object;
 use std::error::Error;
 use tokio_stream::StreamExt;
