+++ src-tauri/src/libs/s3/get/tags.rs
@@ -1,4 +1,4 @@
-use crate::lib::s3::{
+use crate::libs::s3::{
     client::client::create_client,
     utils::response_error::{create_error, ResponseError},
 };
