+++ src-tauri/src/libs/s3/put/files.rs
@@ -1,6 +1,6 @@
-use crate::lib::s3::utils::get_file_name::get_file_name;
-use crate::lib::s3::utils::response_error::create_error;
-use crate::lib::s3::{client::client::create_client, utils::response_error::ResponseError};
+use crate::libs::s3::utils::get_file_name::get_file_name;
+use crate::libs::s3::utils::response_error::create_error;
+use crate::libs::s3::{client::client::create_client, utils::response_error::ResponseError};
 use aws_sdk_s3::primitives::ByteStream;
 use aws_sdk_s3::Client;
 use std::{error::Error, fs, path::Path};
